requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.10
pytest 
pytest-cov 
testcontainers
pyjwt[crypto]==2.12.1